[advanced-reboot] improvements and enable CPU/VLAN ARP watchers during warm reboot (#890)
* [advanced-reboot] move Arista class to seperate module


* [advanced-reboot] use lock to synchronize fast data plane and
reachability_watcher threads


* [advanced-reboot] stabilize test when fast data plane send running

* Apply a filter on socket before sending fast data plane IO
* Save sniffed packets after the traffic test is done


* [advanced-reboot] refactor fast data plane generator code

* reuse from_t1 and from_vlan_server generated packets in
  generate_bidirectional
* use tcp instead ofudp in generate_bidirectional


* [advanced-reboot] add space back
